Move around and clean up components
