+++ client/my-sites/site-settings/date-time-format/style.scss
@@ -1,3 +1,21 @@
+.date-time-format,
+.date-time-format.is-expanded {
+	margin-top: -10px;
+	@include breakpoint( ">480px" ) {
+		margin-top: -16px;
+
+		.foldable-card__header,
+		.foldable-card__content {
+			padding-left: 24px;
+			padding-right: 24px;
+		}
+
+		.foldable-card__action {
+			padding-right: 16px;
+		}
+	}
+}
+
 .date-time-format__custom-field {
 	.form-radio {
 		margin-top: 12px;
